dev(hansbug): add support of dynamic input and output support

+++ pji/service/section/input/link.py
@@ -5,19 +5,20 @@ from typing import Optional, Mapping, Callable
 from hbutils.model import get_repr_info
 from hbutils.string import env_template
 
-from .base import FileInput, FileInputTemplate
+from .base import FileInput, FileInputTemplate, InputCondition, _DEFAULT_INPUT_CONDITION
 from ...base import _check_workdir_file, _check_os_path, _process_environ
 from ....utils import makedirs, wrap_empty
 
 
 class _ILinkFileInput(metaclass=ABCMeta):
-    def __init__(self, file: str, local: str):
+    def __init__(self, file: str, local: str, condition: InputCondition):
         """
         :param file: file path
         :param local: local path
         """
         self.__file = file
         self.__local = local
+        self.__condition = condition
 
     def __repr__(self):
         """
@@ -28,20 +29,22 @@ class _ILinkFileInput(metaclass=ABCMeta):
             args=[
                 ('file', lambda: repr(self.__file)),
                 ('local', lambda: repr(self.__local)),
+                ('condition', lambda: self.__condition.name.lower()),
             ]
         )
 
 
 class LinkFileInputTemplate(FileInputTemplate, _ILinkFileInput):
-    def __init__(self, file: str, local: str):
+    def __init__(self, file: str, local: str, condition=None):
         """
         :param file: file path
         :param local: local path
         """
         self.__file = file
         self.__local = local
+        self.__condition = InputCondition.loads(condition or _DEFAULT_INPUT_CONDITION)
 
-        _ILinkFileInput.__init__(self, self.__file, self.__local)
+        _ILinkFileInput.__init__(self, self.__file, self.__local, self.__condition)
 
     @property
     def file(self) -> str:
@@ -68,20 +71,21 @@ class LinkFileInputTemplate(FileInputTemplate, _ILinkFileInput):
             os.path.abspath(os.path.join(workdir, _check_workdir_file(env_template(self.__local, environ)))))
 
         return LinkFileInput(
-            file=_file, local=_local,
+            file=_file, local=_local, condition=self.__condition,
         )
 
 
 class LinkFileInput(FileInput, _ILinkFileInput):
-    def __init__(self, file: str, local: str, ):
+    def __init__(self, file: str, local: str, condition: InputCondition):
         """
         :param file: file path
         :param local: local path
         """
         self.__file = file
         self.__local = local
+        self.__condition = condition
 
-        _ILinkFileInput.__init__(self, self.__file, self.__local)
+        _ILinkFileInput.__init__(self, self.__file, self.__local, self.__condition)
 
     @property
     def file(self) -> str:
@@ -92,12 +96,18 @@ class LinkFileInput(FileInput, _ILinkFileInput):
         return self.__local
 
     def __call__(self, input_start: Optional[Callable[['LinkFileInput'], None]] = None,
-                 input_complete: Optional[Callable[['LinkFileInput'], None]] = None, **kwargs):
+                 input_complete: Optional[Callable[['LinkFileInput'], None]] = None,
+                 input_skip: Optional[Callable[['LinkFileInput'], None]] = None, **kwargs):
         """
         execute this link event
         """
         wrap_empty(input_start)(self)
-        _parent_path, _ = os.path.split(self.__local)
-        makedirs(_parent_path)
-        os.symlink(self.__file, self.__local, target_is_directory=os.path.isdir(self.__file))
-        wrap_empty(input_complete)(self)
+        if self.__condition == InputCondition.OPTIONAL and not os.path.exists(self.__file):
+            wrap_empty(input_skip)(self)
+        else:
+            if not os.path.exists(self.__file):
+                raise FileNotFoundError(self.__file)
+            _parent_path, _ = os.path.split(self.__local)
+            makedirs(_parent_path)
+            os.symlink(self.__file, self.__local, target_is_directory=os.path.isdir(self.__file))
+            wrap_empty(input_complete)(self)
